pyproject.toml
[project]
name = "marchingCubes"
version = "0.1.0"
description = "Marching Cubes Demo"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "eventlet==0.33.3",
    "flask==2.3.3",
    "flask-socketio==5.3.6",
    "libigl>=2.6.1",
    "numpy>=2.3.3",
    "python-socketio==5.8.0",
    "pythreejs>=2.4.2",
    "scipy>=1.16.2",
    "trimesh==4.0.5",
    "watchdog==3.0.0",
]
